# Implement fast path for small integers (#88)