Performance improvements for random integers with n decimal digits:

MSVC 19.16 x64:
```
n  | Before  | After  | Speed-up
-- | ------- | ------ | --------
1  |  132.6  |  33.5  |  x 3.96
2  |  132.9  |  34.6  |  x 3.84
3  |  124.4  |  34.5  |  x 3.61
4  |  119.8  |  34.9  |  x 3.43
5  |  119.0  |  36.4  |  x 3.27
6  |  113.6  |  37.3  |  x 3.05
7  |  109.0  |  37.7  |  x 2.89
8  |  107.6  |  39.1  |  x 2.75
9  |  103.8  |  40.3  |  x 2.57
10 |  107.1  |  50.4  |  x 2.12
11 |  101.4  |  46.5  |  x 2.18
12 |   96.2  |  47.7  |  x 2.02
13 |   93.8  |  49.0  |  x 1.92
14 |   90.1  |  49.8  |  x 1.81
```

MinGW 8.1 x64:
```
n  | Before  | After  | Speed-up
-- | ------- | ------ | --------
1  |  125.5  |  20.9  |  x 6.02
2  |  122.7  |  20.5  |  x 5.99
3  |  114.7  |  22.5  |  x 5.10
4  |  109.8  |  22.5  |  x 4.89
5  |  112.3  |  23.5  |  x 4.77
6  |  105.0  |  24.1  |  x 4.36
7  |  100.4  |  24.0  |  x 4.19
8  |  100.4  |  25.1  |  x 4.00
9  |   96.2  |  25.1  |  x 3.83
10 |   96.2  |  33.0  |  x 2.92
11 |   94.1  |  28.5  |  x 3.30
12 |   87.9  |  29.8  |  x 2.95
13 |   83.7  |  31.4  |  x 2.67
14 |   81.6  |  32.2  |  x 2.53
```

I have not observed any (significant) degradation in performance for
uniformly distributed random numbers or random bit patterns.