@ulfjack 2019-02-01
We could also special-case small exponents that aren't integers, I think, and that would allow us to remove some of the table entries.